Remove 'application' service, bump pimple/pimple version to 3.2, refactoring commands
